fixed testing/public messaging and tip capturing
testing and public messages were being blocked by the setTimeout,
removed the setTimeout and it worked perfectly.

capturing tips that are public is easy because mfc uses the same grey
background for all non-public (aka ninja) tips. so I check for that
color in the tip before sending the runtime message to the options/games
page